attempting to move socket logic into server index and chat component
