+++ client/socket.js
@@ -1,23 +1,23 @@
-import io from 'socket.io-client'
+// import io from 'socket.io-client'
 
-const socket = io(window.location.origin)
+// const socket = io(window.location.origin)
 
-let messages = [{text: 'hello from client socket.js', id: Math.random()}]
+// let messages = [{text: 'hello from client socket.js', id: Math.random()}]
 
-socket.on('connect', () => {
-  console.log('Connected!', socket.id)
+// socket.on('connect', () => {
+//   console.log('Connected!', socket.id)
 
-  socket.emit('messages', {messages})
+//   socket.emit('messages', {messages})
 
-  socket.on('newMessage', (message) => {
-    console.log('new message')
-    messages = [...messages, message]
-    console.log('newMessages ', messages)
-  })
-  
-})
+//   socket.on('newMessage', (message) => {
+//     console.log('new message')
+//     messages = [...messages, message]
+//     console.log('newMessages ', messages)
+//   })
 
-
+// })
 
 
-export default socket
+
+
+// export default socket
