+++ package.json
@@ -65,6 +65,7 @@
     "redux-thunk": "^2.3.0",
     "sequelize": "^4.38.0",
     "socket.io": "^2.1.1",
+    "socket.io-client": "^2.2.0",
     "twilio": "^3.24.0"
   },
   "devDependencies": {
